LayoutTests/accessibility/mac/figure-element-expected.txt

This test exercises 'figure' element AX properties.

Test figure with aria-label.
figure1 AXRole: AXGroup
figure1 AXRoleDescription: figure
figure1 AXDescription: aria-label attribute
figure1 customContent:
PASS: !titleUIElement === true

Test figure with title attribute.
figure2 AXRole: AXGroup
figure2 AXRoleDescription: figure
figure2 AXDescription: title attribute
figure2 customContent:
PASS: !titleUIElement === true

Test figure with aria-describedby.
figure3 AXRole: AXGroup
figure3 AXRoleDescription: figure
figure3 AXDescription:
figure3 customContent: description: p referenced by aria-describedby
PASS: !titleUIElement === true

Test figure with figcaption element.
figure4 AXRole: AXGroup
figure4 AXRoleDescription: figure
figure4 AXDescription: Figcaption element
figure4 customContent:
PASS: titleUIElement.isEqual(figureCaption) === true

Test figure with figcaption element and title attribute.
figure5 AXRole: AXGroup
figure5 AXRoleDescription: figure
figure5 AXDescription: Figcaption element
figure5 customContent:
PASS: titleUIElement.isEqual(figureCaption) === true

Test display:contents figure with figcaption element.
figure6 AXRole: AXGroup
figure6 AXRoleDescription: figure
figure6 AXDescription: Figcaption element (display:contents)
figure6 customContent:
PASS: titleUIElement.isEqual(figureCaption) === true


PASS successfullyParsed is true

TEST COMPLETE
